// js/src/types/generated/zod/Target.ts
export const TargetID = withCdpMeta(z.string(), "Target.TargetID", "type");
export const SessionID = withCdpMeta(z.string(), "Target.SessionID", "type");
export const TargetInfo = withCdpMeta(z.object({ "targetId": z.lazy(() => TargetID), "type": z.string(), "title": z.string(), "url": z.string(), "attached": z.boolean(), "parentId": z.lazy(() => TargetID).optional(), "openerId": z.lazy(() => TargetID).optional(), "canAccessOpener": z.boolean(), "openerFrameId": z.lazy(() => Page.FrameId).optional(), "parentFrameId": z.lazy(() => Page.FrameId).optional(), "browserContextId": z.lazy(() => Browser.BrowserContextID).optional(), "subtype": z.string().optional() }).passthrough(), "Target.TargetInfo", "type");
export const FilterEntry = withCdpMeta(z.object({ "exclude": z.boolean().optional(), "type": z.string().optional() }).passthrough(), "Target.FilterEntry", "type");
export const TargetFilter = withCdpMeta(z.array(z.lazy(() => FilterEntry)), "Target.TargetFilter", "type");
export const RemoteLocation = withCdpMeta(z.object({ "host": z.string(), "port": z.number().int() }).passthrough(), "Target.RemoteLocation", "type");
export const WindowState = withCdpMeta(z.enum(["normal", "minimized", "maximized", "fullscreen"]), "Target.WindowState", "type");
// ...
export const AttachToTargetParams = withCdpMeta(z.object({ "targetId": z.lazy(() => TargetID), "flatten": z.boolean().optional() }).passthrough(), "Target.attachToTarget.params", "commandParams", { method: "Target.attachToTarget" });
export const AttachToTargetResult = withCdpMeta(z.object({ "sessionId": z.lazy(() => SessionID) }).passthrough(), "Target.attachToTarget.result", "commandResult", { method: "Target.attachToTarget" });
export const AttachToTargetCommand = withCdpCommand("Target.attachToTarget", AttachToTargetParams, AttachToTargetResult);
// ...
export const CreateTargetParams = withCdpMeta(z.object({ "url": z.string(), "left": z.number().int().optional(), "top": z.number().int().optional(), "width": z.number().int().optional(), "height": z.number().int().optional(), "windowState": z.lazy(() => WindowState).optional(), "browserContextId": z.lazy(() => Browser.BrowserContextID).optional(), "enableBeginFrameControl": z.boolean().optional(), "newWindow": z.boolean().optional(), "background": z.boolean().optional(), "forTab": z.boolean().optional(), "hidden": z.boolean().optional(), "focus": z.boolean().optional() }).passthrough(), "Target.createTarget.params", "commandParams", { method: "Target.createTarget" });
export const CreateTargetResult = withCdpMeta(z.object({ "targetId": z.lazy(() => TargetID) }).passthrough(), "Target.createTarget.result", "commandResult", { method: "Target.createTarget" });
export const CreateTargetCommand = withCdpCommand("Target.createTarget", CreateTargetParams, CreateTargetResult);
export const DetachFromTargetParams = withCdpMeta(z.object({ "sessionId": z.lazy(() => SessionID).optional(), "targetId": z.lazy(() => TargetID).optional() }).passthrough(), "Target.detachFromTarget.params", "commandParams", { method: "Target.detachFromTarget" });
export const DetachFromTargetResult = withCdpMeta(z.object({  }).passthrough(), "Target.detachFromTarget.result", "commandResult", { method: "Target.detachFromTarget" });
export const DetachFromTargetCommand = withCdpCommand("Target.detachFromTarget", DetachFromTargetParams, DetachFromTargetResult);
// ...
export const GetTargetsParams = withCdpMeta(z.object({ "filter": z.lazy(() => TargetFilter).optional() }).passthrough(), "Target.getTargets.params", "commandParams", { method: "Target.getTargets" });
export const GetTargetsResult = withCdpMeta(z.object({ "targetInfos": z.array(z.lazy(() => TargetInfo)) }).passthrough(), "Target.getTargets.result", "commandResult", { method: "Target.getTargets" });
export const GetTargetsCommand = withCdpCommand("Target.getTargets", GetTargetsParams, GetTargetsResult);
// ...
export const SetAutoAttachParams = withCdpMeta(z.object({ "autoAttach": z.boolean(), "waitForDebuggerOnStart": z.boolean(), "flatten": z.boolean().optional(), "filter": z.lazy(() => TargetFilter).optional() }).passthrough(), "Target.setAutoAttach.params", "commandParams", { method: "Target.setAutoAttach" });
export const SetAutoAttachResult = withCdpMeta(z.object({  }).passthrough(), "Target.setAutoAttach.result", "commandResult", { method: "Target.setAutoAttach" });
export const SetAutoAttachCommand = withCdpCommand("Target.setAutoAttach", SetAutoAttachParams, SetAutoAttachResult);
// ...
export const SetDiscoverTargetsParams = withCdpMeta(z.object({ "discover": z.boolean(), "filter": z.lazy(() => TargetFilter).optional() }).passthrough(), "Target.setDiscoverTargets.params", "commandParams", { method: "Target.setDiscoverTargets" });
export const SetDiscoverTargetsResult = withCdpMeta(z.object({  }).passthrough(), "Target.setDiscoverTargets.result", "commandResult", { method: "Target.setDiscoverTargets" });
export const SetDiscoverTargetsCommand = withCdpCommand("Target.setDiscoverTargets", SetDiscoverTargetsParams, SetDiscoverTargetsResult);